+++ modules/dsfhub-gcp-mysql/outputs.tf
@@ -1,0 +1,4 @@
+output "this" {
+  description = "GCP MYSQL asset"
+  value       = dsfhub_data_source.this
+}