+++ modules/dsfhub-gcp-pubsub/variables.tf
@@ -20,7 +20,7 @@ variable "audit_pull_enabled" {
 }
 
 variable "audit_type" {
-  description = "Identifier for the type of audit data contained within the PubSub Subscription. Supported values: ALLOYDB_POSTGRESQL, BIGQUERY, BIGTABLE, MYSQL, MYSQL_SLOW, MSSQL, POSTGRESQL, SPANNER."
+  description = "Identifier for the type of audit data contained within the PubSub Subscription. Supported values: ALLOYDB_POSTGRESQL, BIGQUERY, BIGTABLE, GCP_MYSQL_SLOW, MYSQL, MSSQL, POSTGRESQL, SPANNER."
   type        = string
   default     = null
   validation {
@@ -29,8 +29,8 @@ variable "audit_type" {
         "ALLOYDB_POSTGRESQL",
         "BIGQUERY",
         "BIGTABLE",
+        "GCP_MYSQL_SLOW",
         "MYSQL",
-        "MYSQL_SLOW",
         "MSSQL",
         "POSTGRESQL",
         "SPANNER"
@@ -49,6 +49,12 @@ variable "auth_mechanism" {
     condition     = contains(["default", "service_account"], var.auth_mechanism)
     error_message = "Invalid authentication mechanism. Supported values: default, service_account."
   }
+}
+
+variable "content_type" {
+  description = "Desired 'parent' asset 'Server Type' which is the one tha tuses this asset as a destination for logs. NOTE: The content_type field will take precedence on the lookup for parent_asset_id field when checking which server is sending logs to this asset."
+  type        = string
+  default     = null
 }
 
 variable "gateway_id" {